+++ pybamm/models/full_battery_models/lithium_ion/newman_tobias.py
@@ -56,6 +56,25 @@ class NewmanTobias(DFN):
         pybamm.citations.register("Newman1962")
         pybamm.citations.register("Chu2020")
 
+    def set_particle_submodel(self):
+        for domain in ["negative", "positive"]:
+            particle = getattr(self.options, domain)["particle"]
+            phases = ["primary"]
+            for phase in phases:
+                if particle == "Fickian diffusion":
+                    submod = pybamm.particle.FickianDiffusion(
+                        self.param, domain, self.options, phase=phase, x_average=True
+                    )
+                elif particle in [
+                    "uniform profile",
+                    "quadratic profile",
+                    "quartic profile",
+                ]:
+                    submod = pybamm.particle.XAveragedPolynomialProfile(
+                        self.param, domain, self.options, phase=phase
+                    )
+                self.submodels[f"{domain} {phase} particle"] = submod
+
     def set_electrolyte_submodel(self):
 
         surf_form = pybamm.electrolyte_conductivity.surface_potential_form