+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/test_spm.py
@@ -23,6 +23,22 @@ class TestSPM(BaseUnitTestLithiumIon, unittest.TestCase):
         with self.assertRaisesRegex(pybamm.OptionError, "Inverse kinetics"):
             pybamm.lithium_ion.SPM(options)
 
+    def test_x_average_options(self):
+        # Check model with x-averaged side reactions
+        options = {
+            "lithium plating": "irreversible",
+            "lithium plating porosity change": "true",
+            "SEI": "ec reaction limited",
+            "SEI porosity change": "true",
+            "x-average side reactions": "true",
+        }
+        self.check_well_posedness(options)
+
+        # Check model with distributed side reactions throws an error
+        options["x-average side reactions"] = "false"
+        with self.assertRaisesRegex(pybamm.OptionError, "cannot be 'false' for SPM"):
+            pybamm.lithium_ion.SPM(options)
+
     def test_new_model(self):
         model = pybamm.lithium_ion.SPM({"thermal": "x-full"})
         new_model = model.new_copy()
@@ -35,9 +51,13 @@ class TestSPM(BaseUnitTestLithiumIon, unittest.TestCase):
         self.assertEqual(new_model.timescale, model.timescale)
 
         # with custom submodels
-        model = pybamm.lithium_ion.SPM({"thermal": "x-full"}, build=False)
-        particle_n = pybamm.particle.no_distribution.XAveragedPolynomialProfile(
-            model.param, "Negative", "quadratic profile", model.options, "primary"
+        options = {"stress-induced diffusion": "false", "thermal": "x-full"}
+        model = pybamm.lithium_ion.SPM(options, build=False)
+        particle_n = pybamm.particle.XAveragedPolynomialProfile(
+            model.param,
+            "Negative",
+            {**options, "particle": "quadratic profile"},
+            "primary",
         )
         model.submodels["negative primary particle"] = particle_n
         model.build_model()